+++ source/Lib/CommonLib/approx.h
@@ -7,7 +7,7 @@
 
 namespace ApproxSS {
 
-	int __attribute__((optimize("O0"))) start_level(int level = 0); //1 parameters
+	int __attribute__((optimize("O0"))) start_level(int64_t level = 0); //1 parameters
 
 	int __attribute__((optimize("O0"))) end_level(); //0 parameters
